streaming client for Windows

+++ linux/stream.c
@@ -17,13 +17,13 @@ static const unsigned int timeout = 5000;  /* timeout (in ms) for each transfer 
 static atomic_int active_transfers;
 static bool stop_transfers = false;
 /* stream stats */
-static unsigned int success_count = 0;    // number of successful transfers
-static unsigned int failure_count = 0;    // number of failed transfers
-static unsigned long transfer_size = 0;   // total size of data transfers
-static short sample_even_min = SHRT_MAX;  // minimum even sample value
-static short sample_even_max = SHRT_MIN;  // maximum even sample value
-static short sample_odd_min = SHRT_MAX;   // minimum odd sample value
-static short sample_odd_max = SHRT_MIN;   // maximum odd sample value
+static unsigned int success_count = 0;         // number of successful transfers
+static unsigned int failure_count = 0;         // number of failed transfers
+static unsigned long long transfer_size = 0;   // total size of data transfers
+static short sample_even_min = SHRT_MAX;       // minimum even sample value
+static short sample_even_max = SHRT_MIN;       // maximum even sample value
+static short sample_odd_min = SHRT_MAX;        // minimum odd sample value
+static short sample_odd_max = SHRT_MIN;        // maximum odd sample value
 
 static void LIBUSB_CALL transfer_callback(struct libusb_transfer *transfer) ;
 
@@ -138,7 +138,7 @@ void stream_stats(unsigned int duration)
 {
     fprintf(stderr, "success count: %u\n", success_count);
     fprintf(stderr, "failure count: %u\n", failure_count);
-    fprintf(stderr, "transfer size: %lu B\n", transfer_size);
+    fprintf(stderr, "transfer size: %llu B\n", transfer_size);
     fprintf(stderr, "transfer rate: %.0lf kB/s\n", (double) transfer_size / duration / 1024.0);
     fprintf(stderr, "even samples range: [%hd,%hd]\n", sample_even_min, sample_even_max);
     fprintf(stderr, "odd samples range: [%hd,%hd]\n", sample_odd_min, sample_odd_max);
